Move index source package implementations to own files, add .mszyml to content type mapping so that they will be served
